fixed bugs in the lesson 012

+++ script/script.js
@@ -7,7 +7,8 @@ const todoControl = document.querySelector('.todo-control'),
   todoCompleted = document.querySelector('.todo-completed'),
   headerBtn = document.querySelector('.header-button');
 
-const todoData = JSON.parse(localStorage.getItem('toDoList'));;
+let todoData = JSON.parse(localStorage.getItem('toDoList'));
+
 
 
 //JSON.parse(localStorage.getItem('toDoList'));
@@ -23,63 +24,69 @@ const render = () => {
   todoCompleted.textContent = '';
   headerInput.textContent = '';
 
+  if (todoData && todoData) {
+    todoData.forEach((item) => {
 
-  todoData.forEach((item) => {
-    if (!item) {
-      return
-    }
+      const li = document.createElement('li');
 
-    const li = document.createElement('li');
-
-    li.classList.add('todo-item');
-    li.innerHTML = `
+      li.classList.add('todo-item');
+      li.innerHTML = `
       <span class="text-todo">${item.value}</span>
 				<div class="todo-buttons">
 					<button class="todo-remove"></button>
 					<button class="todo-complete"></button>
     `;
-    if (item.completed) {
-      todoCompleted.append(li)
-    } else {
-      todoList.append(li);
-    }
-
-
-
-    const deletedBtn = li.querySelector('.todo-buttons');
-
-
-    deletedBtn.addEventListener('click', (e) => {
-      console.log(e.target)
-      if (e.target.matches('button.todo-complete')) {
-        // console.log(item)
-        completedItem(item);
-      }
-      if (e.target.matches('button.todo-remove')) {
-        // console.log(item)
-        deleteItem(item);
+      if (item.completed) {
+        todoCompleted.append(li)
+      } else {
+        todoList.append(li);
       }
 
 
-      render();
+
+      const deletedBtn = li.querySelector('.todo-buttons');
+
+
+      deletedBtn.addEventListener('click', (e) => {
+        console.log(e.target)
+        if (e.target.matches('button.todo-complete')) {
+          // console.log(item)
+          completedItem(item);
+        }
+        if (e.target.matches('button.todo-remove')) {
+          // console.log(item)
+          deleteItem(item);
+        }
+
+
+        render();
+      })
+
+      const completedItem = (elem) => {
+        console.log(elem)
+        if (elem.value === item.value) {
+          item.completed = !item.completed;
+        }
+      }
+
+      const deleteItem = (elem) => {
+
+        if (elem.value === item.value) {
+          todoData.splice(elem, 1);
+          console.log(todoData.length);
+          if (todoData.length === 0) {
+            localStorage.removeItem('toDoList')
+          }
+        }
+      }
+      addToStorage();
     })
 
-    const completedItem = (elem) => {
-      console.log(elem)
-      if (elem.value === item.value) {
-        item.completed = !item.completed;
-      }
-    }
+  } else {
+    todoData = [];
+  }
 
-    const deleteItem = (elem) => {
 
-      if (elem.value === item.value) {
-        todoData.splice(elem, 1)
-      }
-    }
-
-  })
-  addToStorage();
 };
 
 todoControl.addEventListener('submit', (e) => {
@@ -92,10 +99,11 @@ todoControl.addEventListener('submit', (e) => {
     value: headerInput.value,
     completed: false,
   };
-
+  console.log(todoData)
   todoData.push(newTodo);
   render();
   todoControl.reset();
+  addToStorage();
 });
 
 render();
